requirements.txt
Flask==2.0.3
flask-socketio==5.2.0
psycopg2-binary==2.9.3
eventlet==0.33.1